Divided repository.py into subunits.
